Rename and move InvalidSchemaType exception
As per @tylertreat's suggestion in https://github.com/tylertreat/BigQuery-Python/pull/23
